refactor: 適切な場所へ移動。
