+++ youkai-front/src/components/passwordchallenge/PasswordChallengeProgress.vue
